Renaming client because autorest in js isn't respecting the @client tag
