+++ sdk/healthdataaiservices/azure-health-deidentification/test/public/jobOperationsTest.spec.ts
@@ -8,7 +8,7 @@ import {
   getTestEnvironment,
 } from "./utils/recordedClient.js";
 import { beforeEach, afterEach, it, describe } from "vitest";
-import { DeidServicesClient } from "../../src/clientDefinitions.js";
+import { DeidentificationClient } from "../../src/clientDefinitions.js";
 import { createTestCredential } from "@azure-tools/test-credential";
 import { DeidentificationJob } from "../../src/models.js";
 import { DeidentificationJobOutput, DocumentDetailsOutput } from "../../src/outputModels.js";
@@ -44,7 +44,7 @@ const OUTPUT_FOLDER = "_output";
 
 describe("Batch", () => {
   let recorder: Recorder;
-  let client: DeidServicesClient;
+  let client: DeidentificationClient;
   const environment = getTestEnvironment();
 
   beforeEach(async function (context) {
